+++ src/app/txid/[txId]/redesign/tx-summary/CoinbaseTxSummary.tsx
@@ -1,0 +1,105 @@
+import { AddressLink } from '@/common/components/ExplorerLinks';
+import { BlockLink } from '@/common/components/ExplorerLinks';
+import { formatBlockTime } from '@/common/utils/time-utils';
+import { isConfirmedTx } from '@/common/utils/transaction-utils';
+import { Badge, BlockHeightBadge, DefaultBadgeLabel, TransactionStatusBadge } from '@/ui/Badge';
+
+import {
+  CoinbaseTransaction,
+  MempoolCoinbaseTransaction,
+} from '@stacks/stacks-blockchain-api-types';
+
+import { PriceSummaryItemValue, SummaryItem } from './TxSummary';
+
+export const CoinbaseTxSummaryItems = ({
+  tx,
+}: {
+  tx: CoinbaseTransaction | MempoolCoinbaseTransaction;
+}) => {
+  return (
+    <>
+      <SummaryItem label="ID" value={tx.tx_id} showCopyButton />
+      <SummaryItem
+        label="Status"
+        value={tx.tx_status}
+        valueRenderer={value => <TransactionStatusBadge tx={tx} />}
+        showCopyButton
+      />
+      <SummaryItem
+        label="From"
+        value={tx.sender_address}
+        valueRenderer={value => (
+          <AddressLink principal={value} wordBreak="break-all" variant="tableLink">
+            {value}
+          </AddressLink>
+        )}
+        showCopyButton
+      />
+      {isConfirmedTx<CoinbaseTransaction, MempoolCoinbaseTransaction>(tx) &&
+        tx.coinbase_payload?.alt_recipient && (
+          <SummaryItem
+            label="Reward Recipient"
+            value={tx.coinbase_payload?.alt_recipient}
+            valueRenderer={value => (
+              <AddressLink principal={value} wordBreak="break-all" variant="tableLink">
+                {value}
+              </AddressLink>
+            )}
+            showCopyButton
+          />
+        )}
+      {isConfirmedTx<CoinbaseTransaction, MempoolCoinbaseTransaction>(tx) && (
+        <SummaryItem
+          label="Timestamp"
+          value={formatBlockTime(tx.block_time)}
+          valueRenderer={value => (
+            <Badge variant="solid">
+              <DefaultBadgeLabel label={value} />
+            </Badge>
+          )}
+          showCopyButton
+        />
+      )}
+      <SummaryItem
+        label="Fee"
+        value={tx.fee_rate}
+        valueRenderer={value => <PriceSummaryItemValue value={value} />}
+      />
+      <SummaryItem label="Nonce" value={tx.nonce?.toString() || ''} showCopyButton />
+      {isConfirmedTx<CoinbaseTransaction, MempoolCoinbaseTransaction>(tx) && (
+        <SummaryItem
+          label="Block height"
+          value={tx.block_height?.toString() || ''}
+          showCopyButton
+          valueRenderer={value => <BlockHeightBadge blockType="stx" blockHeight={Number(value)} />}
+        />
+      )}
+      {isConfirmedTx<CoinbaseTransaction, MempoolCoinbaseTransaction>(tx) && (
+        <SummaryItem
+          label="Block hash"
+          value={tx.block_hash?.toString() || ''}
+          showCopyButton
+          valueRenderer={value => (
+            <BlockLink hash={value} wordBreak="break-all">
+              {value}
+            </BlockLink>
+          )}
+        />
+      )}
+      {/* <SummaryItem
+        label="Tenure height"
+        value={tx.tenure_change_payload?.previous_tenure_end || ''}
+        copyable
+        valueRenderer={value => <BlockHeightBadge blockType="stx" blockHeight={Number(value)} />}
+      /> */}
+      {isConfirmedTx<CoinbaseTransaction, MempoolCoinbaseTransaction>(tx) && (
+        <SummaryItem
+          label="Bitcoin Anchor"
+          value={tx.burn_block_height?.toString() || ''}
+          showCopyButton
+          valueRenderer={value => <BlockHeightBadge blockType="btc" blockHeight={Number(value)} />}
+        />
+      )}
+    </>
+  );
+};
